requirements.txt
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fonttools==4.61.1
fuzzywuzzy==0.18.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.17.0
nest-asyncio==1.6.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.2
plotly==6.6.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
seaborn==0.13.2
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
tomli==2.4.0
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
unicode==2.9
Unidecode==1.4.0
urllib3==2.6.3
wcwidth==0.6.0
